move verification implementation to did
