+++ pallets/attestation/src/benchmarking.rs
@@ -16,10 +16,11 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use delegation::{benchmarking::setup_delegations, Permissions, VerifyDelegateSignature};
+use delegation::{benchmarking::setup_delegations, Permissions};
 use frame_benchmarking::{benchmarks, impl_benchmark_test_suite};
 use frame_support::traits::{Currency, Get};
 use frame_system::RawOrigin;
+use kilt_support::signature::VerifyDelegateSignature;
 use sp_core::sr25519;
 use sp_runtime::traits::Hash;
 use sp_std::num::NonZeroU32;
