+++ runtimes/spiritnet/Cargo.toml
@@ -19,6 +19,7 @@ pallet-vesting = {git = "https://github.com/paritytech/substrate", default-featu
 # KILT pallets & primitives
 kilt-launch = {path = "../../pallets/kilt-launch", default-features = false}
 kilt-primitives = {path = "../../primitives", default-features = false}
+kilt-support = {path = "../../support", default-features = false, optional = true}
 parachain-staking = {default-features = false, path = "../../pallets/parachain-staking"}
 
 # Substrate dependencies
@@ -93,6 +94,8 @@ runtime-benchmarks = [
   "frame-system-benchmarking",
   "hex-literal",
   "kilt-launch/runtime-benchmarks",
+  "kilt-support",
+  "kilt-support/runtime-benchmarks",
   "pallet-balances/runtime-benchmarks",
   "pallet-collective/runtime-benchmarks",
   "pallet-democracy/runtime-benchmarks",
