+++ support/Cargo.toml
@@ -4,7 +4,7 @@ description = "Shared traits and structs used across the KILT pallets"
 edition = "2018"
 name = "kilt-support"
 repository = "https://github.com/KILTprotocol/mashnet-node"
-version = "0.26.0"
+version = "0.26.1"
 
 [dependencies]
 codec = {default-features = false, features = ["derive"], package = "parity-scale-codec", version = "2.0.0"}
@@ -17,12 +17,17 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [features]
 default = ["std"]
-try-runtime = [
-  "frame-support/try-runtime"
+mock = []
+runtime-benchmarks = [
+  "sp-runtime/runtime-benchmarks",
+  "frame-support/runtime-benchmarks",
 ]
 std = [
   "codec/std",
   "frame-support/std",
   "sp-runtime/std",
-  "sp-std/std"
+  "sp-std/std",
+]
+try-runtime = [
+  "frame-support/try-runtime",
 ]
